+++ src/main/java/com/mango/amango/domain/auth/service/ReissueTokenService.java
@@ -1,6 +1,6 @@
 package com.mango.amango.domain.auth.service;
 
-import com.mango.amango.domain.auth.entity.dto.response.TokenRes;
+import com.mango.amango.domain.auth.presentation.dto.response.TokenRes;
 import jakarta.servlet.http.HttpServletRequest;
 
 public interface ReissueTokenService {
