+++ src/main/java/com/mango/amango/domain/auth/service/SignInService.java
@@ -1,7 +1,7 @@
 package com.mango.amango.domain.auth.service;
 
-import com.mango.amango.domain.auth.entity.dto.request.LoginReq;
-import com.mango.amango.domain.auth.entity.dto.response.TokenRes;
+import com.mango.amango.domain.auth.presentation.dto.request.LoginReq;
+import com.mango.amango.domain.auth.presentation.dto.response.TokenRes;
 import jakarta.servlet.http.HttpServletResponse;
 
 public interface SignInService {
