+++ src/main/java/com/mango/amango/domain/auth/service/impl/SignInServiceImpl.java
@@ -1,7 +1,7 @@
 package com.mango.amango.domain.auth.service.impl;
 
-import com.mango.amango.domain.auth.entity.dto.request.LoginReq;
-import com.mango.amango.domain.auth.entity.dto.response.TokenRes;
+import com.mango.amango.domain.auth.presentation.dto.request.LoginReq;
+import com.mango.amango.domain.auth.presentation.dto.response.TokenRes;
 import com.mango.amango.domain.auth.service.SignInService;
 import com.mango.amango.domain.user.entity.User;
 import com.mango.amango.domain.user.service.UserService;
